Add new po_directory fixture
This fixture is test-scoped and copies the contents of the
translations directory created by PootleEnv into a tempdir for
use by other fixtures, and sets the POOTLE_TRANSLATIONS_DIRECTORY
setting.

It also cleans up afterwards.

This replaces the old fixture of the same name
